presentations/test-the-web-forward-overview/index.html

<!doctype html>  
<html lang="en">
	
	<head>
		<meta charset="utf-8">
		
		<title>Test the Web Forward</title>
        <meta name="apple-mobile-web-app-capable" content="yes" />
		<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
		
		<link href='http://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
		
		<link rel="stylesheet" href="css/reset.css">
		<link rel="stylesheet" href="css/main.css">
		<link rel="stylesheet" href="css/print.css" type="text/css" media="print">

		<link rel="stylesheet" href="lib/textmate.css">
		<link rel="stylesheet" href="css/adobe.css">
	</head>
	
	<body>
		
		<div class="reveal">

			<!-- Used to fade in a background when a specific slide state is reached -->
			<div class="state-background"></div>
			
			<!-- Any section element inside of this container is displayed as a slide -->
			<div class="slides">
				<section>
					<h2>Test the Web Forward</h2>
					<img src="assets/images/W3C_logo.jpg">
					<script>
						// Delicously hacky. Look away.
						if( navigator.userAgent.match( /(iPhone|iPad|iPod|Android)/i ) ) document.write( '<p style="color: rgba(0,0,0,0.3); text-shadow: none;">('+'Tap to navigate'+')</p>' );
					</script>
				</section>
				
				<section>
					<h2>Reftest Overview</h2>
					<blockquote>
					<ul>
					    <li class="fragment">Two files: Test file + Reference file</li>
					    <li class="fragment">Test file uses the feature you're testing</li>
					    <li class="fragment">Reference file is an exact visual match without using the feature you're testing</li>
						<li class="fragment">Self-describing (works both manually and with automation)</li>
						<li class="fragment"> Cross-browser & cross-platform</li>
					</ul>
					</blockquote>
				</section>
				
				<section>
    				<h2>Directory structure</h2>
    				<section>
    				    <br>
    				    
    				    <ul>
    				    <li style="list-style-type: none;">CSSWG Test Repository</li>
    				    <ul>
    				        <li><tt>contributors</tt></li>
						    <ul>
						        <li><tt>ttwf</tt></li>
						        <ul>
						            <li><tt><em>your_csswg_id</em></tt></li>
						            <ul>
						                <li><tt>incoming</tt></li>
    						            <li><tt>submitted</tt></li>
						            </ul>
						        </ul>
						    </ul>
						</ul>
					
					</section>
					<section>
					    <img src="assets/images/directoryStructure.png"/>
					</section>
    			</section>
    			<section>
    				<h2>Sample Test Case</h2>
    				<section>
    					<br><br>
    					<p style="font-size: 28pt">Basic test for the transform property described in the CSS3 Transforms spec using the <code>translate()</code> function</p>
    					<br>
    					<br>
    					<p>Spec Links:</p>
    					<p style="font-size: 16pt">
    					    <a href="http://www.w3.org/TR/css3-transforms/css3-transforms/#transform-property" target="top">http://www.w3.org/TR/css3-transforms/css3-transforms/#transform-property</a>
    					    <br>
    					    <a href="http://www.w3.org/TR/css3-transforms/css3-transforms/#two-d-transform-functions" target="top">http://www.w3.org/TR/css3-transforms/css3-transforms/#two-d-transform-functions</a>
					    </p>
					</section>
					<section>
					    <img src="assets/images/specAssertion.png">
					</section>
	            </section>
				
				<section data-state="screenshotPage">
				    <br>
				    <h3>Sample files</h3>
    			    <img src="assets/images/001a.png">
    			    <img src="assets/images/001b.png">
            	</section>
			
				<section data-state="expandPageWide">
				    <br>
				    <h3>Required Metadata: Title</h3>
				    <br>
				    <p>Test file</p>
				    <pre>
					<code class="html">
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
    &lt;head&gt;
    &lt;title&gt;CSS Transforms Test: transform property with translate function&lt;/title&gt;
        . . .			
    </span>		        
				    </code>
				    </pre>
		            <p>Reference file</p>
		            <pre>
                    <code>
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
    &lt;head&gt;
    &lt;title&gt;CSS Transforms Reference file&lt;/title&gt;
        . . .			
</span>
                    </code>
                    </pre>
			    </section>
                
                <section  data-state="expandPageWide">
                    <br>
    			    <h3>Required Metadata: Author</h3>
                    <br>
                    <p>Test file</p>
                    <pre>
                    <code class="html">
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    . . .			
    &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
    . . .			
</span>		        
				    </code>
				    </pre>
		            <p>Reference file</p>
                    <pre>
                    <code>
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    . . .
    &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
    . . .			
</span>
                    </code>
                    </pre>
                </section>
                
                <section  data-state="expandPageWide">
                    <br>
    			    <h3>Required Metadata: Spec Links</h3>
    			    <p>Test file only</p>
                    <pre>
                    <code>
<span style="font-size: 16pt;" class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
  . . .
  &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
  &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;          
  . . .
</span>
                    </code>
                    </pre>
    			</section>

                <section  data-state="expandPageWide">
                    <br>
    			    <h3>Required Metadata: Reference File Path</h3>
        		    <p>Test file only</p>
                    <pre>
                    <code>
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    . . .   
    &lt;link rel="match" href="reference/ttwf-reftest-tutorial-ref.html"&gt;
    . . .
</span>
                        </code>
                        </pre>
    			</section>
    		    
    		    <section data-state="expandPageWide">
    		        <br>
    		        <h3>Required Metadata: Assertion</h3>
                    <section data-state="expandPageWide">
            	        <p>Test file only</p>
                        <pre>
                        <code>
<span style="font-size: 16pt;" class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
. . .   
&lt;meta name="assert" 
  content="This transform moves the element by 100 pixels in both the X and Y directions."&gt;
. . .
</span>
                        </code>
                        </pre>
    			    </section>
    			    <section data-state="expandPageWide">
			            <img src="assets/images/specAssertion.png">
			        </section>
    			</section>
    		
    			
    			<section  data-state="expandPageWide">
    			    <br>
    			    <h3>The Test</h3>
    			    <section data-state="expandPageWide">
    				    <p>Test element with the transform applied</p>
    				    <pre>
    				    <code class="html">
<span class="expand2">
                &lt;<span class="styleTag">style type</span>="text/css"&gt;
                    <span class="styleTag">.greenSquare</span> {
                        <span class="cssProperty">position: absolute;
                        top: 0;
                        left: 0;
                        width: 100<span class="px">px</span>;
                        height: 100<span class="px">px</span>;
                        background: green;</span>
                        transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);
                    }
                    . . .
                &lt;/<span class="styleTag">style</span>&gt;</span>				        
    				    </code>
    				    </pre>
    			    </section>

    			    <section  data-state="expandPageWide">
    				    <p>Reference element</p>
    				    <pre>
    				    <code class="html">
<span class="expand2">
                    &lt;<span class="styleTag">style type</span>="text/css"&gt;
                        <span class="styleTag">.greenSquare</span> {
                            <span class="cssProperty">position: absolute;
                            top: 100;
                            left: 100;
                            width: 100<span class="px">px</span>;
                            height: 100<span class="px">px</span>;
                            background: green;</span>
                        }
                    &lt;/<span class="styleTag">style</span>&gt;
</span>				        
    				    </code>
    				    </pre>
    			    </section>
                </section>
                
                <section  data-state="expandPageWide">
                    <br>
    			    <h3>Visible failure</h3>
    			    <section data-state="expandPageWide">
    			        <p>Element is only visible when the test fails</p>
    			         <pre>
        				    <code class="html">
<span class="expand2">
                        &lt;<span class="styleTag">style type</span>="text/css"&gt;
                            . . .
                            <span class="styleTag">.redSquare</span> {
                                <span class="cssProperty">position: absolute;
                                top: 101<span class="px">px</span>;
                                left: 101<span class="px">px</span>;
                                width: 98<span class="px">px</span>;
                                height: 98<span class="px">px</span>;
                                background: red;</span>
                            }
                        &lt;/<span class="styleTag">style</span>&gt;    
</span>				        
        				    </code>
        				    </pre>
    			    </section>
    			    <section data-state="expandPageWide">
    				    <p>Failing test example</p>
    				    <img src="assets/images/failingRefTest.png"/>
    			    </section>
    			</section>
    			
    			<section data-state="expandPageWide">
    			    <section data-state="expandPageWide">
    			    <br>
    			    <h3>Self Description</h3>
    			    <p style="font-size: 24pt">Simple statement of how the page renders when the test passes</p>
    				<br>
    				<p style="font-size: 24pt"> Test file & Reference file </p>
    				 <pre>
    				 <code class="html">
<span class="expand2">
    &lt;body&gt;
        &lt;p&gt;The test passes if there is a green square and no red.&lt;/p&gt;
        . . .
    &lt;/body&gt;
</span>    				        
    			    </code>
    			    </pre>
                </section>
                </section>
                
                <section  data-state="expandPageWide">
                    <section data-state="expandPageWide">
        		    <h3>Add the elements</h3>
        		    <p>Test file</p>
        		     <pre>
    				 <code class="html">
<span class="expand2">
                &lt;body&gt;
                    . . .
                    &lt;div class="redSquare"&gt;&lt;/div&gt;
                    &lt;div class="greenSquare"&gt;&lt;/div&gt;
                &lt;/body&gt;
</span>    				        
    			    </code>
    			    </pre>
    			    <p>Reference file</p>
                     <pre>
    				 <code class="html">
<span class="expand2">
                &lt;body&gt;
                    . . .
                    &lt;div class="greenSquare"&gt;&lt;/div&gt;
                &lt;/body&gt;
</span>    				        
    			    </code>
    			    </pre>
                </section>
                </section>
                
                <section data-state="expandPageWide">
                    <br>
                    <h2>Vendor prefixes</h2>
               
                   <section data-state="expandPageWide">
                       <br>
                       <blockquote>
                       <ul>
                           <li>You may need to add a vendor prefix to the CSS property you're testing</li>
                           <li>If the prefix is necessary, your test will appear to fail without it</li>
                           <li>However, do not include vendor prefixes when pushing tests to the W3C</li>
                       </ul>   
                       </blockquote> 
                       <br>
                       
                  </section>
              
                  <section data-state="expandPageWide">
                      <pre>
                      <code class="html">
                        <span class="expand2">
                <span class="styleTag">.greenSquare</span> {
                    <span class="cssProperty">position: absolute;
                    top: 0;
                    left: 0;
                    width: 100<span class="px">px</span>;
                    height: 100<span class="px">px</span>;
                    <span class="highlight">-webkit-transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                    <span class="highlight">-moz-transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                    <span class="highlight">-ms-transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                    <span class="highlight">-o-transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                    background: green;
                }</span> 
                      </code>
                      </pre>
                  </section>
                      
                  </section>
    			
				<!-- Begin JS TEST -->

				<section>
					<h2>JavaScript Test Overview</h2>
					<br>
					<ul>
					    <li>JavaScript tests programmatically verify functionality</li>
					    <li>JavaScript tests have some advantages over Ref Tests
					        <ul>
					            <li>Robust</li>
					            <li>Flexible</li>
					            <li>Performant</li>
					            <li>Powerful</li>
					        </ul>
					    </li>
					    <li>Ref Tests have some advantages over JavaScript tests
					        <ul>
					            <li>Verify complete rendering</li>
					            <li>Easy to understand</li>
					            <li>Works for clients that don't support JavaScript</li>
					            <li>Avoid introducing additional technologies (JS)</li>
					        </ul>
					    </li>
					</ul>
			    </section>
			    
				<section>
				    <h2>JavaScript Test Overview (cont'd)</h2>
				    <br>
                    <ul>
    					<li>JavaScript can be used in conjunction with Ref Tests
    					    <ul>
    				            <li>Test can have both a reference file and a JavaScript section</li>
    				            <li>Ref Test portion supports clients without script engines</li>
    				        </ul>
    				    </li>
    				    <li>Self-describing RefTests are the preferred test format</li>
    				    <li>JavaScript tests should be used in the following scenarios:
    				        <ul>
    				            <li>Testing a JS API or behavior (ex: CSSOM spec)</li>
    				            <li>Adding automation to Ref Tests</li>
    				        </ul>
    				    </li>
    				    <li>W3C provides a framework (<tt>testharness.js</tt>) to simplify and standardize JS test creation</li>
					</ul>
				</section>	    


    				<section>
    			        <h3>testharness</h3>
    			        <p><tt>testharness.js</tt> and <tt>testharnessreport.js</tt></p>
    			        <pre>
    			            <code class="html">
    &lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;

    &lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;
                            </code>
                        </pre>    
    				    <ul>
    				        <li>JavaScript API for making common test assertions</li>
    				        <li>Should be used by all JavaScript tests</li>
    				        <li>Path must be to <tt>/resources</tt> directory at root level</li>
    				        <li>Element with <code>id="log"</code> should exist in test file
    				            <ul>
    				                <li>Test results table will be added to this element, if it exists</li>
    				            </ul>
    				        </li>
    				    </ul>
    				</section>

    				<section>
    			        <section>
        			        <h2>The test function</h2>
            			    <br>
            			    <pre>
            			        <code class="html">
            test(test_function, name, properties)
            			        </code>
            			    </pre>
            			    <ul>
            			        <li>Used for synchronous tests
            			            <ul>
            			                <li><code>async_test()</code> is also provided by testharness for asychronous testing</li>
            			            </ul>
            			        </li>
            			        <li>Documented and implemented in <tt>testharness.js</tt></li>
            			        <li>Used to define an individual test case in a file
            			            <ul>
            			                <li>Common for a file to have multiple calls to <code>test()</code></li>
            			            </ul>
            			        </li>   
            			        <li>testharness will display PASS/FAIL for each call to <code>test()</code></li>

            			    </ul>
                        </section>

                        <section>
                            <h2>test() Argument 1: test_function</h2>
            			    <pre>
            			        <code class="html">
            test(<span style="color:red">test_function</span>, name, properties)
            			        </code>
            			    </pre>                        
                            <ul>
                                <li><code>test_function</code> must be an object, not a function call</li>
                                <li><code>test_function</code> will typically be one of the assert methods supplied by testharness
                                    <ul>
                                        <li>Can also be user defined if no assert method provides the needed functionality</li>
                                    </ul>
                                </li>
                                <li>Result from the assertion will determine PASS/FAIL</li>
                                <li>Usage example:
                                    <pre style="font-size:18px">
                                        <code>
            test(function() {assert_true(true)}, name, properties)
                                        </code>
                                    </pre>  
                                </li>                            
        			        </ul>                   
                        </section>

                       <section>
                            <h2>test() Argument 2: Name</h2>
                            <pre>
                                <code class="html">
            test(test_function,<span style="color:red"> name</span>, properties)
                                </code>
                            </pre>                        
                            <ul>
                                <li><code>name</code> is a string that identifies the test</li> 
                                <li><code>name</code> should be short, unique, and must not change</li>
            			        <li><code>name</code> will be displayed in the test results table</li>
                                <li>Usage example:
                                    <pre style="font-size:18px">
                                        <code class="html">
        test(function() {assert_true(true)}, "test_name", properties)
                                        </code>
                                    </pre>  
                                </li>
                            </ul>       			                                   
                        </section>

                        <section>
                            <h2>test() Argument 3: Properties</h2>
                            <pre>
                                <code class="html">
                test(test_function, name,<span style="color:red"> properties</span>)
                                </code>
                            </pre>                        
                            <ul>
                                <li><code>properties</code> is an object that overrides defaults</li>
                                <li>Recognized properties include <code>timeout</code> and metadata</li>
                                <li>Usage example:
                                    <pre style="font-size:18px">
                                        <code class="html">
 test(function() {assert_true(true)}, "test_name", <br><br> {timeout:1000, <br><br> help: 'http://www.w3.org/TR/spec#section', <br><br> author: ['Bill Gates &lt;bgates@msn.com&gt;', 'Steve Jobs http://apple.com/sjobs'],<br><br> assert: 'This should always be true.'})
                                        </code>
                                    </pre>                                
                                </li>    			                    
        			        </ul>
                        </section>
                        
                        <section>
                            <h2>Per-test Metadata</h2>
                           <pre style="font-size:18px">
                               <code class="html">
test(function() {assert_true(true)}, "test_name", <br><br>{ help: 'http://www.w3.org/TR/spec#section', <br><br>author: ['Bill Gates &lt;bgates@msn.com&gt;', 'Steve Jobs http://apple.com/sjobs'],<br><br>assert: 'This test something.', 'This also tests something else.'})
                               </code>
                           </pre>
                            <ul>
                                <li>Used when a file contains multiple calls to <code>test()</code></li>
                                <li>Only needed when an individual test's metadata is different than what is defined in <code>head</code></li>
                                <li><code>name</code> passed to <code>test()</code> should be short, unique, and must not change</li>
                                <li><code>assert</code> metadata should be used for descriptive info</li>
                            </ul>
                        </section>
                        
                    </section>

            	    <section>
            	        <section>
        			        <h2>Assert methods</h2>
            			    <ul>
            			        <li><tt>testharness.js</tt> provides multiple assert methods
            			            <ul>
            			                <li>Full list is provided in the comment section of <tt>testharness.js</tt></li>
            			                <li>Typically supplied as the <code>test_function</code> for <code>test()</code></li>
            			            </ul>
            			        </li>
            			        <li>All assert methods contain an optional <code>description</code> argument that is only output if the assertion fails
            			            <ul>
            			                <li>Used to provide additional debug information</li>
            			            </ul>
            			        </li>
            			        <li>All asserts must be located in a call to <code>test()</code>
            			            <ul>
            			                <li>asserts outside won't be detected correctly by the harness and may cause a file to stop testing</li>
            			            </ul>
            			        </li>
            			    </ul>
        			    </section>

        			    <section>
        			        <h2>Assert Example</h2>
                            <pre>
                                <code class="html">
        assert_equals(actual, expected, description)
                                </code>
                            </pre>
                            <ul>
                                <li><code>actual</code>: The actual value from the functionality being tested</li>
                                <li><code>expected</code>: The expected value</li>
                                <li><code>description</code> (optional): Output only if the assertion fails</li>
                                <li>Usage example:
                                    <pre style="font-size:18px">
                                        <code class="html">
 assert_equals(getActualData("myElement"), 100, "Expected value for myElement is 100")
                                        </code>
                                    </pre>                                
                                </li>                            
                            </ul>    			        
        			    </section>
    				</section>

                	<section>
        				<h2>Sample Test Case</h2>
        				<section>
        					<br>
        					<p>API test for the transform property described in the CSS3 Transforms spec using the <code>translate()</code> function</p>
        					<br><br>
        					<p>Spec Links:</p>
        					<p style="font-size: 16pt">
        					    <a href="http://www.w3.org/TR/css3-transforms/css3-transforms/#transform-property" target="top">http://www.w3.org/TR/css3-transforms/css3-transforms/#transform-property</a>
        					    <br>
        					    <a href="http://www.w3.org/TR/css3-transforms/css3-transforms/#two-d-transform-functions" target="top">http://www.w3.org/TR/css3-transforms/css3-transforms/#two-d-transform-functions</a>
    					    </p>
    					</section>
    					<section>
    					    <img src="assets/images/specAssertion.png">
    					</section>
    	            </section>			

    			    <section>
    			        <h3>Metadata</h3>
    			        <section>
        			        <pre style="width:975px">
        			            <code class="html" style="line-height:26px; font-size:18px">
&lt;head&gt;
  &lt;title&gt;CSS Transforms API Test: transform translate&lt;/title&gt;
  &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
  &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
  &lt;link rel="help" 
    href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;
  &lt;meta name="flags" content="dom"&gt;
  &lt;meta name="assert" content="The transform should be translate(100px,100px)"&gt;	
...    
&lt;/head&gt;                
        			            </code>
        			        </pre>
        			        <ul style="line-height:60px">
        			            <li>JavaScript tests have similar requirements as RefTest</li>
        			            <li><code>dom</code> token in <code>flags</code> metadata specifies that JavaScript is required</li>
        			            <li>JSON metadata supports build system (not shown)</li>
        			    </section>
        			    <section>
        			        <pre>
        			            <code class="html" style="line-height:26px; font-size:24px">
&lt;head&gt;
...    
    &lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
    &lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;    
&lt;/head&gt;    			                
    			            </code>
        			        </pre>
        			        <ul>
        			            <li><tt>testharness.js</tt> and <tt>testharnessreport.js</tt> imported</li>
        			            <li>Path to both scripts points at <code>/resources</code> directory at root level
        			                <ul>
        			                    <li>This is the required path before submitting to CSSWG Test Repository</li>
        			                </ul>
        			            </li>
        			        </ul>
        			    </section>
    			    </section>				

			    
			    <section data-state="largePage">
			        <h3>Body</h3>
			        <pre style="width:900px">
			            <code class="html" style="line-height:22px; font-size:14px">
&lt;body&gt;
    &lt;div id="myDiv"&gt;&lt;/div&gt;
    &lt;div id="log"&gt;&lt;/div&gt;    

    &lt;script&gt;
        // Set the transform 
        document.getElementById("myDiv").style.transform = "translate(100px,100px)";

        // Verify that the transform was set as expected
        test(function() {assert_equals(
                document.getElementById("myDiv").style.getPropertyCSSValue("transform").cssText, // Actual
                "translate(100px, 100px)",  // Expected
                "transform should be translate(100px,100px)")}, // Description
                "Transform_Translate_100px_100px"); // name
    &lt;/script&gt;
&lt;/body&gt;			                
			            </code>
			        </pre>
			        <ul style="font-size:28px">
			            <li><code>myDiv</code> div used as test element - translate applied in JS</li>
			            <li><code>log</code> div used as container for test results</li>
			            <li><code>test()</code> uses <code>assert_equals</code> as <code>test_function</code></li>
                        <li></code>assert_equals</code> actual argument supplied by API call</li>
			        </ul>			        
			    </section>
 			    
    			<section data-state="largePage">
    			    <section>
    				    <h3>Test Results</h3>
    				    <p>PASS</p>
    				    <img src="assets/images/jsTest_PASS.png"/>
    			    </section>
    			    
                    <section>
    				    <h3>Test Results</h3>
    				    <p>FAIL</p>
    				    <img src="assets/images/jsTest_FAIL.png"/>
    			    </section>    			    
                </section>                                                    
    			
    			<section>
    			    <h1>Questions & Answers</h1>
    			</section>
		
			</div>

			<!-- The navigational controls UI -->
			<aside class="controls">
				<a class="left" href="#">&#x25C4;</a>
				<a class="right" href="#">&#x25BA;</a>
				<a class="up" href="#">&#x25B2;</a>
				<a class="down" href="#">&#x25BC;</a>
			</aside>

			<!-- Displays presentation progress, max value changes via JS to reflect # of slides -->
			<div class="progress"><span></span></div>
			
		</div>

		<!-- Optional libraries for code syntax highlighting and classList support in IE9 -->
		<script src="lib/highlight.js"></script>
		<script src="lib/classList.js"></script>
		
		<script src="js/reveal.js"></script>
		
		<script>
			// Parse the query string into a key/value object
			var query = {};
			location.search.replace( /[A-Z0-9]+?=(\w*)/gi, function(a) {
				query[ a.split( '=' ).shift() ] = a.split( '=' ).pop();
			} );

			// Fires when a slide with data-state=customevent is activated
			Reveal.addEventListener( 'customevent', function() {
				alert( '"customevent" has fired' );
			} );

			// Fires each time a new slide is activated
			Reveal.addEventListener( 'slidechanged', function( event ) {
				// event.indexh & event.indexv
			} );

			Reveal.initialize({
				// Display controls in the bottom right corner
				controls: true,

				// Display a presentation progress bar
				progress: true,

				// If true; each slide will be pushed to the browser history
				history: true,

				// Loops the presentation, defaults to false
				loop: false,

				// Flags if mouse wheel navigation should be enabled
				mouseWheel: true,

				// Apply a 3D roll to links on hover
				rollingLinks: true,

				// UI style
				theme: query.theme || 'default', // default/neon

				// Transition style
				transition: query.transition || 'default' // default/cube/page/concave/linear(2d)
			});

			hljs.initHighlightingOnLoad();
		</script>

	</body>
</html>
